refactor: CrawlingModeController now manages Dungeon information. Removed DungeonController and replaced with DungeonBuilder.
