+++ Solution/DungeonCrawler/Player/Player.cs
@@ -2,7 +2,6 @@ namespace CaptainCoder.Dungeoneering.Player;
 
 using CaptainCoder.Dungeoneering.DungeonMap;
 
-[Serializable]
 public record PlayerView(Position Position, Facing Facing)
 {
     public PlayerView(int x, int y, Facing facing) : this(new Position(x, y), facing) { }
